✨ #4 rootにErrorBoundary実装
