--- app/routes/$.tsx
@@ -1,37 +1,0 @@
-import { isRouteErrorResponse, useRouteError } from "@remix-run/react";
-
-// loaderで意図的に`404`をthrowして、エラーUIを描画する
-// https://remix.run/docs/en/main/route/error-boundary
-export async function loader() {
-  throw new Response("Not found", { status: 404 });
-}
-
-export default function NotFound() {
-  return <ErrorBoundary />;
-}
-
-export function ErrorBoundary() {
-  const error = useRouteError();
-
-  if (isRouteErrorResponse(error)) {
-    return (
-      <div>
-        <h1>
-          {error.status} {error.statusText}
-        </h1>
-        <p>{error.data}</p>
-      </div>
-    );
-  } else if (error instanceof Error) {
-    return (
-      <div>
-        <h1>Error</h1>
-        <p>{error.message}</p>
-        <p>The stack trace is:</p>
-        <pre>{error.stack}</pre>
-      </div>
-    );
-  } else {
-    return <h1>Unknown Error</h1>;
-  }
-}
